SqlConnector: Improve collect data handling
We make use of the new *OrNull types available in ginger and
also use the new iterator support in AbstractCollection to
provide the traversable PDOStatement instead of the populated
resultset. This allows processing of much larger resultsets
as long as we are in the same process and collection is not
serialized.